Moving attribute renderer and configured microdata into the CurationConcerns::Renderer namespace. Refers to #766
Remove byebug point
